central/requirements.txt
AMQPStorm==2.7.2
asyncio==3.4.3
attrs==19.3.0
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
clickclick==1.2.2
configparser==4.0.2
connexion==2.4.0
cryptography==2.8
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.1
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.21.0
multidict==4.6.1
mysql-connector-python==8.0.18
openapi-spec-validator==0.2.8
pamqp==2.3.0
protobuf==3.11.1
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.6
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
swagger-ui-bundle==0.0.6
urllib3==1.25.7
Werkzeug==0.16.0
yarl==1.3.0
